no more namespace std

+++ ZAPD/StringHelper.h
@@ -2,8 +2,8 @@
 
 #include <numeric>
 #include <stdarg.h>
-#include <string.h>
 #include <string>
+#include <cstring>
 #include <vector>
 
 class StringHelper
@@ -66,7 +66,7 @@ public:
 
 	static bool EndsWith(const std::string& s, const std::string& input)
 	{
-		int32_t inputLen = strlen(input.c_str());
+		size_t inputLen = strlen(input.c_str());
 		return s.rfind(input) == (s.size() - inputLen);
 	}
